Match NumPy regarding .npy for names in .npz files
Closes #48.

When writing a .npz file, NumPy unconditionally adds ".npy" to every
name. When generating the public list of names when reading a .npz
file, NumPy strips a single ".npy" (if present) from each name. When
accessing an array in a .npz file by name, it first checks if that
exact name is in the .npz file, and if not, it then tries the name
with ".npy" appended. While I personally dislike the
inconsistency/implicitness of this behavior, ndarray-npy should follow
it for compatibility with NumPy.